+++ util.go
@@ -13,6 +13,7 @@ import (
 	"mime/multipart"
 	"net/http"
 	"os"
+	"path"
 	"path/filepath"
 	"reflect"
 	"runtime"
@@ -182,4 +183,30 @@ func releaseBuffer(buf *bytes.Buffer) {
 		buf.Reset()
 		bufPool.Put(buf)
 	}
+}
+
+func composeRequestURL(pathURL string, c *Client, r *Request) string {
+	if !strings.HasPrefix(pathURL, "/") {
+		pathURL = "/" + pathURL
+	}
+
+	reqURL := "/"
+	for _, segment := range strings.Split(pathURL, "/") {
+		if strings.HasPrefix(segment, "{") && strings.HasSuffix(segment, "}") {
+			key := segment[1 : len(segment)-1]
+			if val, found := r.pathParams[key]; found {
+				reqURL = path.Join(reqURL, val)
+				continue
+			}
+
+			if val, found := c.pathParams[key]; found {
+				reqURL = path.Join(reqURL, val)
+				continue
+			}
+		}
+
+		reqURL = path.Join(reqURL, segment)
+	}
+
+	return reqURL
 }
